+++ src/visualstudio/TechLeadTools.VisualStudio/Properties/AssemblyInfo.cs
@@ -7,5 +7,5 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyProduct("TechLeadTools")]
 [assembly: AssemblyCopyright("Copyright © Zinfo 2026")]
 [assembly: ComVisible(false)]
-[assembly: AssemblyVersion("0.1.2.0")]
-[assembly: AssemblyFileVersion("0.1.2.0")]
+[assembly: AssemblyVersion("0.1.3.0")]
+[assembly: AssemblyFileVersion("0.1.3.0")]